Add country flags to periods list
