+++ internal/model/session.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"os"
 	"strings"
+	"telegram-tax-bot/internal/utils"
 )
 
 type Session struct {
@@ -58,7 +59,13 @@ func (s *Session) BuildPeriodsList() string {
 		if out == "" {
 			out = "по " + s.Data.Current
 		}
-		builder.WriteString(fmt.Sprintf("%d. %s — %s (%s)\n", i+1, in, out, p.Country))
+		flag := ""
+		if p.Country == "unknown" {
+			flag = "🕳 "
+		} else if code, ok := utils.CountryCodeMap[p.Country]; ok {
+			flag = utils.CountryToFlag(code) + " "
+		}
+		builder.WriteString(fmt.Sprintf("%d. %s — %s (%s%s)\n", i+1, in, out, flag, p.Country))
 	}
 	return builder.String()
 }
